+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   + Add details in validation error messages when using `min` and/or `max`.
   + `v-lunaris` directive supports radio inputs.
   + Add a tip when fallback to _id when using function getPrimaryKey().
+  + Improve debug information for lunaris-vue plugin.
 - Fix :
   + Change vue component hook to remove hooks.
   + Empty arrays were not considarated as correct values in cache. As a consequence, HTTP requests were fired.